Added exceptions for book and user and fixed services

+++ Infrastructure/Services/AuthorService.cs
@@ -1,9 +1,11 @@
 ﻿using Abstractions.Repositories;
 using Application.DTOs.Authors;
+using Application.DTOs.Books;
 using Application.Services.Interfaces;
 using AutoMapper;
 using Domain.Entities;
 using Domain.Exceptions;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,11 +17,13 @@ namespace Infrastructure.Services
     public class AuthorService : IAuthorService
     {
         private readonly IAuthorRepository _authorRepository;
+        private readonly ILogger<AuthorService> _logger;
         private readonly IMapper _mapper;
-        public AuthorService(IAuthorRepository authorRepository,IMapper mapper)
+        public AuthorService(IAuthorRepository authorRepository, IMapper mapper, ILogger<AuthorService> logger)
         {
             _authorRepository = authorRepository;
             _mapper = mapper;
+            _logger = logger;
         }
         public async Task<GetAuthorDTO> AddAsync(CreateAuthorDTO createAuthorDTO)
         {
@@ -27,62 +31,100 @@ namespace Infrastructure.Services
                 throw new ArgumentNullException(nameof(createAuthorDTO));
 
             var authorToCreate = _mapper.Map<Author>(createAuthorDTO);
-
             var existingAuthor = await _authorRepository.GetByIdAsync(authorToCreate.Id);
 
             if (existingAuthor != null)
             {
-
+                _logger.LogWarning($"Author {createAuthorDTO.FirstName} {createAuthorDTO.Surname} already exists");
                 throw new AuthorExistsException(createAuthorDTO.FirstName, createAuthorDTO.Surname);
             }
 
             var author = await _authorRepository.AddAsync(authorToCreate);
 
-            if(author == null)
+            if (author == null)
+            {
+                _logger.LogWarning($"Failed to create author with name {createAuthorDTO.FirstName} {createAuthorDTO.Surname}");
                 throw new InvalidOperationException("Failed to create author");
+            }
 
+            _logger.LogInformation($"Successfully added author with id {author.Id}");
             return _mapper.Map<GetAuthorDTO>(author);
         }
         public async Task<IEnumerable<GetAuthorDTO>> GetAllAsync()
         {
             var authors = await _authorRepository.GetAllAsync();
 
+            if (authors == null || authors.Any())
+            {
+                _logger.LogInformation("No authors found in repository");
+                return Enumerable.Empty<GetAuthorDTO>();
+            }
+
             return _mapper.Map<IEnumerable<GetAuthorDTO>>(authors);
         }
         public async Task<GetAuthorDTO> GetByIdAsync(Guid id)
         {
+            if (id == Guid.Empty)
+                throw new ArgumentNullException(nameof(id));
+
             var author = await _authorRepository.GetByIdAsync(id);
 
             if (author == null)
+            {
+                _logger.LogInformation($"Author with id {id} not found");
                 throw new AuthorNotFoundException(id);
+            }
 
             return _mapper.Map<GetAuthorDTO>(author);
         }
         public async Task<GetAuthorDTO> UpdateAsync(UpdateAuthorDTO updateAuthorDTO)
         {
+            if (updateAuthorDTO == null)
+                throw new ArgumentNullException(nameof(updateAuthorDTO));
+
             var authorExists = await _authorRepository.GetByIdAsync(updateAuthorDTO.Id);
 
             if (authorExists == null)
+            {
+                _logger.LogInformation($"Author with id {updateAuthorDTO.Id} not found");
                 throw new AuthorNotFoundException(updateAuthorDTO.Id);
+            }
 
             _mapper.Map(updateAuthorDTO, authorExists);
 
             var author = await _authorRepository.UpdateAsync(authorExists);
+            if (author == null)
+            {
+                _logger.LogWarning($"Failed to update author with id {authorExists.Id}");
+                throw new InvalidOperationException($"Failed to update author with id {authorExists.Id}");
+            }
 
             return _mapper.Map<GetAuthorDTO>(authorExists);
         }
         public async Task<bool> DeleteAsync(Guid id)
         {
+            if (id == Guid.Empty)
+                throw new ArgumentNullException(nameof(id));
+
             var author = await _authorRepository.GetByIdAsync(id);
 
             if (author == null)
+            {
+                _logger.LogInformation($"Author with id {id} not found");
                 throw new AuthorNotFoundException(id);
+            }
 
-            await _authorRepository.DeleteAsync(author.Id);
+            var result = await _authorRepository.DeleteAsync(author.Id);
 
+            if (!result)
+            {
+                _logger.LogWarning($"Failed to delete author with id {id}");
+                throw new InvalidOperationException("Failed to delete author");
+            }
+
+            _logger.LogInformation("Successfully deleted author with ID: {authorId}", id);
             return true;
         }
-
 
         public async Task<GetAuthorDTO> GetByFullNameAsync(string surname)
         {
@@ -92,7 +134,11 @@ namespace Infrastructure.Services
             var author = await _authorRepository.GetByFullNameAsync(surname);
 
             if (author == null)
+            {
+                _logger.LogInformation($"Author with surname {surname} not found");
                 throw new AuthorNotFoundException(surname);
+
+            }
 
             return _mapper.Map<GetAuthorDTO>(author);
         }
